finish home page, add totalStock count in product, fix recommend products add home page product count in yaml

+++ src/main/java/com/sarkhan/backend/model/product/items/Color.java
@@ -16,9 +16,9 @@ public class Color {
 
     int photoCount;
 
-    Double stock;
+    Long stock;
 
     List<String> images;
 
-    Map<String, Double> sizeStockMap;
+    Map<String, Long> sizeStockMap;
 }
